Use akracun's tmux and vitality.vim
This makes it so I don't have to run in bracketed paste mode always in
order to get cmd+v pasting to work in vim. Now it'll just work in vim.

This also enables FocusGained and FocusLost events in tmux, so you can do with
those as you please.

Of course, none of this works w/o building and using akracun's tmux.